modified requirements
For the time of Python 3 transition, its recommended to clone the parent
repository from:

  https://github.com/return42/moz-cloud

With parent 'moz-cloud' (and for the time of transition) we are able to install
requirements from relative paths, e.g.::

  -e file:../m2crypto#egg=M2Crypto --global-option="build_ext"

Using relative paths and installing requirements from the 'moz-cloud' helps to
speed up development. Later (before upstream) relative pathes should be removed.

Signed-off-by: Markus Heiser <markus.heiser@darmarit.de>